test(vetting): fold the leftover "remaining" fixture header into the morphology one
test_2d_remaining_common.h was the fixture half of the pre-SPEC test_2d_remaining_features.h
bucket; Wave 6 split that file by family and the name outlived the grouping it described. Its
five fixtures move into test_2d_morphology_common.h, now the only shared 2D shape fixture header,
and the two half-fixtures over the same shape2d ROI become one pass -- they recomputed basic
morphology, contour, hull and erosion each, and their settings builders were byte-for-byte
identical. Feature order is a common supersequence of both, so every pinned golden is unchanged:
783/783 gtest cases pass.

Names the assert helpers for the tables they read (SPEC 6.3.1), dropping the pre-SPEC
unvetted_no_direct_oracle / verifiable_with_3p_builtin_oracle vocabulary. Inside a _regression
file both mean regression, and the latter hinted at a third-party claim SPEC 1 says is not there;
the SCOPED_TRACE labels follow, matching the three helpers in the file that already read
REGRESSION__.

Stops two comments from citing helpers that no longer exist -- one named the shared lookup and
the shared assert helper that the value split removed, the other named an imea helper whose name
now belongs to a different, live one in test_2d_morphology_imea.h.

Repoints the docs that named the deleted header, corrects a config recipe still pointing at
make_remaining2d_settings(), de-duplicates the not_covered.md fixture row with its counts, and
drops a stale note in test_2d_morphology_cellprofiler.h claiming its goldens come from a shared
map in a _common.h -- they have been local since the split.
