+++ tests/test_2d_zernike_regression.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "test_2d_remaining_common.h"
+#include "test_2d_morphology_common.h"
 
 static ref_vals_map<std::vector<double>> zernike_2d_regression_ref_vals{
 	{"ZERNIKE2D", {
@@ -36,13 +36,13 @@ static void assert_zernike_vector_feature_regression(
 // ---------------------------------------------------------------------------------------------------
 // Migrated from test_2d_remaining_features.h (Wave 6): ZERNIKE2D. Per registry decision (§6.1) mahotas
 // is not an accepted oracle, so ZERNIKE2D stays analytic/regression -> test_2d_zernike_regression.h.
-// Shared fixture/oracle-data lives in test_2d_remaining_common.h.
+// Shared fixture/oracle-data lives in test_2d_morphology_common.h.
 // ---------------------------------------------------------------------------------------------------
 
 void test_2d_zernike_moments_regression()
 {
 	std::vector<std::vector<double>> fvals;
-	calculate_remaining2d_shape_feature_values(fvals);
+	calculate_shape2d_feature_values(fvals);
 
 	assert_zernike_vector_feature_regression(fvals, Nyxus::Feature2D::ZERNIKE2D, "ZERNIKE2D");
 }
