+++ tests/vetting/not_covered.md
@@ -17,9 +17,9 @@ Regenerate by re-deriving both sets; keep this file in step with each reorg wave
 
 ---
 
-## A. Test files no registry row references — 25 files, 58 test functions
+## A. Test files no registry row references — 24 files, 58 test functions
 
-### A.1 Correctly absent — plumbing, fixtures and framework self-tests (18 files)
+### A.1 Correctly absent — plumbing, fixtures and framework self-tests (17 files)
 
 These assert no feature value, so they have no `(feature × config × oracle)` row by construction
 (SPEC §1). Recording them here so their absence is a documented decision, not an oversight.
@@ -37,7 +37,7 @@ These assert no feature value, so they have no `(feature × config × oracle)` r
 | `test_2d_ooc_invariant.py` | 9 | out-of-core == in-RAM equality; spans all features, per-feature rows would be meaningless |
 | `test_2d_ooc_mechanics.py` | 1 | oversized-montage failure path |
 | `test_vetting_mechanics.py` | 5 | self-test of `check_coverage.py` |
-| `test_3d_coverage_common.h`, `test_3d_morphology_common.h`, `test_2d_moments_common.h`, `test_2d_morphology_common.h`, `test_2d_remaining_common.h` | 0 | shared fixtures; the kind belongs to the files that include them |
+| `test_3d_coverage_common.h`, `test_3d_morphology_common.h`, `test_2d_moments_common.h`, `test_2d_morphology_common.h` | 0 | shared fixtures; the kind belongs to the files that include them |
 | `test_feature_calculation_common.h` | 0 | shared `test_feature` template helper |
 
 ### A.2 Assert feature values but no row lists them — real gaps (7 files)
